init: first commit

+++ spuddy-security-objects/src/main/java/io/github/buzzxu/spuddy/security/objects/Organization.java
@@ -91,7 +91,7 @@ public class Organization extends Id<Integer> implements Tree<Integer> {
 
     @Override
     public int id() {
-        return id.intValue();
+        return id;
     }
 
     @Override
